# perf(engine): Parallelise reading inputs of SortMerge