--- a/pkg/dataobj/sections/logs/table_test.go
+++ b/pkg/dataobj/sections/logs/table_test.go
@@ -44,21 +44,22 @@ func initBuffer(buf *tableBuffer) {
 func Test_mergeTables(t *testing.T) {
 	var buf tableBuffer
 
+	// tables need to be sorted by Timestamp DESC and StreamID ASC
 	var (
 		tableA = buildTable(&buf, 1024, dataset.CompressionOptions{}, []Record{
-			{StreamID: 1, Timestamp: time.Unix(1, 0), Line: []byte("hello")},
-			{StreamID: 2, Timestamp: time.Unix(2, 0), Line: []byte("are")},
-			{StreamID: 3, Timestamp: time.Unix(3, 0), Line: []byte("goodbye")},
+			{StreamID: 3, Timestamp: time.Unix(3, 0), Line: []byte("hello")},
+			{StreamID: 2, Timestamp: time.Unix(2, 0), Line: []byte("how")},
+			{StreamID: 1, Timestamp: time.Unix(1, 0), Line: []byte("you")},
 		})
 
 		tableB = buildTable(&buf, 1024, dataset.CompressionOptions{}, []Record{
 			{StreamID: 1, Timestamp: time.Unix(2, 0), Line: []byte("world")},
-			{StreamID: 3, Timestamp: time.Unix(1, 0), Line: []byte("you")},
+			{StreamID: 3, Timestamp: time.Unix(1, 0), Line: []byte("goodbye")},
 		})
 
 		tableC = buildTable(&buf, 1024, dataset.CompressionOptions{}, []Record{
-			{StreamID: 2, Timestamp: time.Unix(1, 0), Line: []byte("how")},
-			{StreamID: 3, Timestamp: time.Unix(2, 0), Line: []byte("doing?")},
+			{StreamID: 3, Timestamp: time.Unix(2, 0), Line: []byte("are")},
+			{StreamID: 2, Timestamp: time.Unix(1, 0), Line: []byte("doing?")},
 		})
 	)