--- a/pkg/engine/executor/sortmerge.go
+++ b/pkg/engine/executor/sortmerge.go
@@ -5,9 +5,11 @@ import (
 	"fmt"
 	"slices"
 	"sort"
+	"sync"
 
 	"github.com/apache/arrow-go/v18/arrow"
 	"github.com/apache/arrow-go/v18/arrow/array"
+	"golang.org/x/sync/errgroup"
 
 	"github.com/grafana/loki/v3/pkg/engine/planner/physical"
 )
@@ -113,49 +115,62 @@ start:
 	timestamps := make([]int64, 0, len(p.inputs))
 	inputIndexes := make([]int, 0, len(p.inputs))
 
-loop:
+	var wg errgroup.Group
+	var mtx sync.Mutex
+
 	for i := range len(p.inputs) {
 		// Skip exhausted inputs
 		if p.exhausted[i] {
-			continue loop
+			continue
 		}
 
-		// Load next batch if it hasn't been loaded yet, or if current one is already fully consumed
-		// Read another batch as long as the input yields zero-length batches.
-		for p.batches[i] == nil || p.offsets[i] == p.batches[i].NumRows() {
-			// Reset offset
-			p.offsets[i] = 0
+		wg.Go(func() error {
+			// Load next batch if it hasn't been loaded yet, or if current one is already fully consumed
+			// Read another batch as long as the input yields zero-length batches.
+			for p.batches[i] == nil || p.offsets[i] == p.batches[i].NumRows() {
+				// Reset offset
+				p.offsets[i] = 0
 
-			// Read from input
-			err := p.inputs[i].Read()
-			if err != nil {
-				if errors.Is(err, EOF) {
-					p.exhausted[i] = true
-					p.batches[i] = nil // remove reference to arrow.Record from slice
-					continue loop
+				// Read from input
+				err := p.inputs[i].Read()
+				if err != nil {
+					if errors.Is(err, EOF) {
+						p.exhausted[i] = true
+						p.batches[i] = nil // remove reference to arrow.Record from slice
+						return nil
+					}
+					return err
 				}
-				return err
+
+				// It is safe to use the value from the Value() call, because the error is already checked after the Read() call.
+				// In case the input is exhausted (reached EOF), the return value is `nil`, however, since the flag `p.exhausted[i]` is set, the value will never be read.
+				p.batches[i], _ = p.inputs[i].Value()
 			}
 
-			// It is safe to use the value from the Value() call, because the error is already checked after the Read() call.
-			// In case the input is exhausted (reached EOF), the return value is `nil`, however, since the flag `p.exhausted[i]` is set, the value will never be read.
-			p.batches[i], _ = p.inputs[i].Value()
-		}
+			// Fetch timestamp value at current offset
+			col, err := p.columnEval(p.batches[i])
+			if err != nil {
+				return err
+			}
+			tsCol, ok := col.ToArray().(*array.Timestamp)
+			if !ok {
+				return errors.New("column is not a timestamp column")
+			}
+			ts := tsCol.Value(int(p.offsets[i]))
 
-		// Fetch timestamp value at current offset
-		col, err := p.columnEval(p.batches[i])
-		if err != nil {
-			return err
-		}
-		tsCol, ok := col.ToArray().(*array.Timestamp)
-		if !ok {
-			return errors.New("column is not a timestamp column")
-		}
-		ts := tsCol.Value(int(p.offsets[i]))
+			mtx.Lock()
+			// Populate slices for sorting
+			inputIndexes = append(inputIndexes, i)
+			timestamps = append(timestamps, int64(ts))
+			mtx.Unlock()
 
-		// Populate slices for sorting
-		inputIndexes = append(inputIndexes, i)
-		timestamps = append(timestamps, int64(ts))
+			return nil
+		})
+
+	}
+
+	if err := wg.Wait(); err != nil {
+		return err
 	}
 
 	// Pipeline is exhausted if no more input batches are available