--- a/pkg/querier/querier_test.go
+++ b/pkg/querier/querier_test.go
@@ -90,6 +90,9 @@ func TestQuerier_Label_QueryTimeoutConfigFlag(t *testing.T) {
 func mockQuerierConfig() Config {
 	return Config{
 		TailMaxDuration: 1 * time.Minute,
+		Engine: logql.EngineOpts{
+			BatchSize: 1,
+		},
 	}
 }