### Summary

This PR changes how inputs of the `SortMerge` operator are consumed. While on the base branch `chaudum/sort-dataobj-sections-by-timestamp-desc` the inputs are read sequentially, this branch reads them in parallel.

### Benchmark comparison to base branch

```console
$ benchstat logql_bench_sortbytimestamp.txt logql_bench_sortbytimestamp_parallel.txt
goos: linux
goarch: amd64
pkg: github.com/grafana/loki/v3/pkg/logql/bench
cpu: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
                                                                                                                                                              │ logql_bench_sortbytimestamp.txt │ logql_bench_sortbytimestamp_parallel.txt │
                                                                                                                                                              │             sec/op              │      sec/op       vs base                │
LogQL/query={region="ap-southeast-1",_env="dev"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                        776.6m ± 12%       396.6m ± 24%  -48.94% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1",_env="dev"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                              1577.0m ± 27%       898.1m ± 15%  -43.05% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1",_env="dev"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                 1.709 ±  3%        1.299 ± 33%  -23.99% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1",_env="dev"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                           1.548 ± 32%        1.163 ± 10%  -24.89% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                                  1377.1m ±  8%       811.5m ± 22%  -41.07% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                         1234.0m ± 41%       959.6m ± 15%  -22.23% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                            1.200 ±  2%        1.412 ± 21%        ~ (p=0.143 n=10)
LogQL/query={region="ap-southeast-1"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                      1.649 ± 29%        1.074 ± 12%  -34.85% (p=0.001 n=10)
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                     1.769 ± 27%        1.253 ±  6%  -29.15% (p=0.000 n=10)
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                            3.576 ± 13%        3.012 ± 17%  -15.78% (p=0.005 n=10)
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                             2.565 ±  1%        2.838 ± 10%  +10.64% (p=0.007 n=10)
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                       3.538 ± 14%        2.866 ± 19%  -19.00% (p=0.005 n=10)
LogQL/query={service_name="kubernetes"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                                1207.5m ±  6%       758.2m ±  5%  -37.21% (p=0.000 n=10)
LogQL/query={service_name="kubernetes"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                         1.577 ± 32%        1.334 ± 18%        ~ (p=0.218 n=10)
LogQL/query={service_name="kubernetes"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                          1.008 ±  7%        1.274 ±  8%  +26.41% (p=0.000 n=10)
LogQL/query={service_name="kubernetes"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                  1409.2m ± 29%       548.2m ± 71%  -61.10% (p=0.000 n=10)
geomean                                                                                                                                                                             1.591              1.178        -25.97%
```